+++ src/detection/initsystem/initsystem_windows.c
@@ -25,7 +25,7 @@ const char* ffDetectInitSystem(FFInitSystemResult* result) {
     if (ffIsSystemBasicProcessInfoAvailable()) {
         // SYSTEM_BASICPROCESS_INFORMATION entries are much smaller than SYSTEM_PROCESS_INFORMATION ones,
         // so a modest buffer should be enough to contain all processes
-        SYSTEM_BASICPROCESS_INFORMATION buffer[1024];
+        SYSTEM_BASICPROCESS_INFORMATION buffer[1024] = {};
         NTSTATUS status = NtQuerySystemInformation(SystemBasicProcessInformation, buffer, sizeof(buffer), NULL);
         if (status != STATUS_INFO_LENGTH_MISMATCH && !NT_SUCCESS(status)) {
             goto fallback;